Fixes to sell.php and bugs relating to the 0,00 format
git-svn-id: svn://svn.code.sf.net/p/simpleauction/svn/WeBid/trunk@285 cb2008ec-e0ba-457b-a47b-93945ab2bf72

+++ includes/functions_sell.inc.php
@@ -37,9 +37,9 @@ function setvars()
 
 	$with_reserve = (isset($_POST['with_reserve'])) ? $_POST['with_reserve'] : $_SESSION['SELL_with_reserve'];
 	$reserve_price = (isset($_POST['reserve_price'])) ? $_POST['reserve_price'] : $_SESSION['SELL_reserve_price'];
-	$minimum_bid = (int)(isset($_POST['minimum_bid'])) ? $_POST['minimum_bid'] : $_SESSION['SELL_minimum_bid'];
+	$minimum_bid = (isset($_POST['minimum_bid'])) ? $_POST['minimum_bid'] : $_SESSION['SELL_minimum_bid'];
 	$minimum_bid = (empty($minimum_bid)) ? 0.99 : $minimum_bid;
-	$shipping_cost = (int)(isset($_POST['shipping_cost'])) ? $_POST['shipping_cost'] : $_SESSION['SELL_shipping_cost'];
+	$shipping_cost = (isset($_POST['shipping_cost'])) ? $_POST['shipping_cost'] : $_SESSION['SELL_shipping_cost'];
 	$shipping_cost = (empty($shipping_cost)) ? 0 : $shipping_cost;
 	$imgtype = (isset($_POST['imgtype'])) ? $_POST['imgtype'] : $_SESSION['SELL_file_uploaded'];
 	$title = (isset($_POST['title'])) ? $_POST['title'] : $_SESSION['SELL_title'];
